+++ app/client/src/interfaces/Payments.ts
@@ -1,0 +1,10 @@
+export interface RepaymentData {
+    days: number;
+    date: Date;
+    principalDue: number;
+    loanBalance: number;
+    interest: number;
+    fees: number;
+    penalties: number;
+    due: number;
+}
