+++ app/client/src/interfaces/Stakes.ts
@@ -1,0 +1,6 @@
+export interface StakerMetadata {
+    img: string;
+    name: string;
+    relationship: string;
+    verifiedIds: string[];
+}
