+++ app/client/src/interfaces/Tokens.ts
@@ -1,0 +1,5 @@
+export interface TokenMetadata {
+    imgSrc: string;
+    name: string;
+    address: string;
+}
