+++ app/client/src/interfaces/index.ts
@@ -1,0 +1,5 @@
+export * from './Loans';
+export * from './Payments';
+export * from './Stakes';
+export * from './Tokens';
+export * from './Users';
